Move get_codeblock_node method to util.py
Decided to move this method since it was to
be used in more than one place.
